pyproject.toml
[project]
name = "update-workflows"
version = "0.1"
description = "Update GitHub Actions Workflows"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
update-workflows = "update_workflows:main"

[tool.setuptools]
py-modules = ["update_workflows"]
